refactor(core): extract formatApiError into @eventuras/core/errors
Moves the Problem-Details + SDK-error-shape parser out of apps/web and
into @eventuras/core/errors as a framework-neutral helper so every
server action (and any future consumer, including non-Next.js ones)
can surface meaningful API failures through the same code path.

While moving, fix two gaps flagged in review of the original copy:

- Plain string error bodies (the SDK's fallback for non-JSON responses)
  now pass through as the error message instead of being replaced with
  the generic fallback.
- Problem Details `detail` field is now consulted before falling back
  to `title` / legacy shapes, so non-validation errors no longer lose
  the backend's explanation.

Events' updateEvent now imports the shared helper. Registrations'
updateRegistration and patchRegistration gained matching error
surfacing — previously they returned a flat "Failed to update
registration" string regardless of what the API actually said.
