+++ apps/web/src/app/(admin)/admin/events/actions.ts
@@ -3,6 +3,7 @@
 import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
 
+import { formatApiError } from '@eventuras/core/errors';
 import {
   actionError,
   actionSuccess,
@@ -152,39 +153,6 @@ const updateLogger = Logger.create({
   context: { module: 'action:updateEvent' },
 });
 
-/**
- * Extract a readable message from the SDK's `error` object. Handles ASP.NET
- * Problem Details (RFC 7807) validation errors — `{ errors: { Field: [msgs] } }`
- * — by surfacing per-field issues instead of the generic "Bad Request".
- */
-function formatApiError(raw: unknown): string {
-  const fallback = 'An error occurred while updating the event';
-  if (!raw || typeof raw !== 'object') return fallback;
-
-  const err = raw as {
-    errors?: Record<string, string[] | string | undefined>;
-    title?: string;
-    body?: { message?: string };
-    message?: string;
-    statusText?: string;
-  };
-
-  if (err.errors && typeof err.errors === 'object') {
-    const parts = Object.entries(err.errors)
-      .map(([field, msgs]) => {
-        const text = Array.isArray(msgs) ? msgs.join(', ') : msgs;
-        return text ? `${field}: ${text}` : field;
-      })
-      .filter(Boolean);
-    if (parts.length > 0) {
-      const prefix = err.title ? `${err.title} — ` : '';
-      return `${prefix}${parts.join('; ')}`;
-    }
-  }
-
-  return err.body?.message || err.message || err.statusText || fallback;
-}
-
 export async function updateEvent(
   eventId: number,
   eventData: EventFormDto
@@ -245,7 +213,10 @@ export async function updateEvent(
 
       updateLogger.error({ eventId, ...errorDetails }, errorMsg);
 
-      const userMessage = formatApiError(response.error);
+      const userMessage = formatApiError(
+        response.error,
+        'An error occurred while updating the event'
+      );
       return actionError(userMessage, 'API_ERROR', errorDetails);
     }
