+++ libs/core/src/errors/index.ts
@@ -1,0 +1,1 @@
+export { formatApiError } from './formatApiError.js';
