+++ libs/core/vite.config.ts
@@ -6,5 +6,6 @@ export default defineVanillaLibConfig({
     'datetime/index': 'src/datetime/index.ts',
     'currency/index': 'src/currency/index.ts',
     'url/index': 'src/url/index.ts',
+    'errors/index': 'src/errors/index.ts',
   },
 });
